Add finished epam project

+++ utils/src/main/java/utils/StringUtils.java
@@ -2,7 +2,7 @@ package utils;
 
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
- 	if (Integer.parseInt(str)>0) {
+ 	if (Double.parseDouble(str)>0) {
     		return true;
     	} else return false;
     }
